+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/db/AppDatabase.kt
@@ -3,10 +3,13 @@ package au.com.shiftyjelly.pocketcasts.models.db
 import android.content.Context
 import android.database.sqlite.SQLiteDatabase
 import androidx.core.content.contentValuesOf
+import androidx.room.AutoMigration
 import androidx.room.Database
+import androidx.room.DeleteColumn
 import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import androidx.room.migration.AutoMigrationSpec
 import androidx.room.migration.Migration
 import androidx.sqlite.db.SupportSQLiteDatabase
 import au.com.shiftyjelly.pocketcasts.model.BuildConfig
@@ -63,8 +66,11 @@ import au.com.shiftyjelly.pocketcasts.localization.R as LR
         UserEpisode::class,
         PodcastRatings::class,
     ],
-    version = 79,
+    version = 83,
     exportSchema = true,
+    autoMigrations = [
+        AutoMigration(from = 81, to = 82, spec = AppDatabase.Companion.DeleteSilenceRemovedMigration::class),
+    ],
 )
 @TypeConverters(
     AnonymousBumpStat.CustomEventPropsTypeConverter::class,
@@ -493,6 +499,69 @@ abstract class AppDatabase : RoomDatabase() {
             )
         }
 
+        val MIGRATION_79_80 = addMigration(79, 80) { database ->
+            database.execSQL(
+                """
+                    ALTER TABLE podcasts
+                    ADD COLUMN start_from_modified INTEGER
+                """.trimIndent(),
+            )
+            database.execSQL(
+                """
+                    ALTER TABLE podcasts
+                    ADD COLUMN skip_last_modified INTEGER
+                """.trimIndent(),
+            )
+        }
+
+        val MIGRATION_80_81 = addMigration(80, 81) { database ->
+            database.execSQL(
+                """
+                    ALTER TABLE podcasts
+                    ADD COLUMN auto_add_to_up_next_modified INTEGER
+                """.trimIndent(),
+            )
+            database.execSQL(
+                """
+                    ALTER TABLE podcasts
+                    ADD COLUMN override_global_effects_modified INTEGER
+                """.trimIndent(),
+            )
+            database.execSQL(
+                """
+                    ALTER TABLE podcasts
+                    ADD COLUMN playback_speed_modified INTEGER
+                """.trimIndent(),
+            )
+            database.execSQL(
+                """
+                    ALTER TABLE podcasts
+                    ADD COLUMN volume_boosted_modified INTEGER
+                """.trimIndent(),
+            )
+            database.execSQL(
+                """
+                    ALTER TABLE podcasts
+                    ADD COLUMN trim_silence_level_modified INTEGER
+                """.trimIndent(),
+            )
+        }
+
+        @DeleteColumn(
+            tableName = "podcasts",
+            columnName = "silence_removed",
+        )
+        class DeleteSilenceRemovedMigration : AutoMigrationSpec
+
+        val MIGRATION_82_83 = addMigration(82, 83) { database ->
+            database.execSQL(
+                """
+                    ALTER TABLE podcasts
+                    ADD COLUMN show_notifications_modified INTEGER
+                """.trimIndent(),
+            )
+        }
+
         fun addMigrations(databaseBuilder: Builder<AppDatabase>, context: Context) {
             databaseBuilder.addMigrations(
                 addMigration(1, 2) { },
@@ -862,6 +931,10 @@ abstract class AppDatabase : RoomDatabase() {
                 MIGRATION_76_77,
                 MIGRATION_77_78,
                 MIGRATION_78_79,
+                MIGRATION_79_80,
+                MIGRATION_80_81,
+                // 81 to 82 added via auto migration
+                MIGRATION_82_83,
             )
         }