+++ modules/services/model/src/main/java/au/com/shiftyjelly/pocketcasts/models/to/StatsBundle.kt
@@ -2,6 +2,8 @@ package au.com.shiftyjelly.pocketcasts.models.to
 
 import java.util.Date
 
+// Prefer using StatsBundleData instead of StatsBundle. Eventually,
+// it would be nice to remove StatsBundle entirely.
 data class StatsBundle(
     val values: Map<String, Long>,
     val startedAt: Date?,