+++ modules/services/servers/src/main/java/au/com/shiftyjelly/pocketcasts/servers/sync/NamedSettings.kt
@@ -19,12 +19,36 @@ data class NamedSettingsSettings(
 data class ChangedNamedSettings(
     @field:Json(name = "skipForward") val skipForward: NamedChangedSettingInt? = null,
     @field:Json(name = "skipBack") val skipBack: NamedChangedSettingInt? = null,
+    @field:Json(name = "gridLayout") val gridLayout: NamedChangedSettingInt? = null,
     @field:Json(name = "gridOrder") val gridOrder: NamedChangedSettingInt? = null,
     @field:Json(name = "marketingOptIn") val marketingOptIn: NamedChangedSettingBool? = null,
     @field:Json(name = "freeGiftAcknowledgement") val freeGiftAcknowledgement: NamedChangedSettingBool? = null,
     @field:Json(name = "autoArchivePlayed") val autoArchiveAfterPlaying: NamedChangedSettingInt? = null,
     @field:Json(name = "autoArchiveInactive") val autoArchiveInactive: NamedChangedSettingInt? = null,
     @field:Json(name = "autoArchiveIncludesStarred") val autoArchiveIncludesStarred: NamedChangedSettingBool? = null,
+    @field:Json(name = "playbackSpeed") val playbackSpeed: NamedChangedSettingDouble? = null,
+    @field:Json(name = "trimSilence") val trimSilence: NamedChangedSettingInt? = null,
+    @field:Json(name = "volumeBoost") val volumeBoost: NamedChangedSettingBool? = null,
+    @field:Json(name = "rowAction") val rowAction: NamedChangedSettingInt? = null,
+    @field:Json(name = "upNextSwipe") val upNextSwipe: NamedChangedSettingInt? = null,
+    @field:Json(name = "mediaActions") val showCustomMediaActions: NamedChangedSettingBool? = null,
+    @field:Json(name = "mediaActionsOrder") val mediaActionsOrder: NamedChangedSettingString? = null,
+    @field:Json(name = "episodeGrouping") val episodeGrouping: NamedChangedSettingInt? = null,
+    @field:Json(name = "keepScreenAwake") val keepScreenAwake: NamedChangedSettingBool? = null,
+    @field:Json(name = "openPlayer") val openPlayerAutomatically: NamedChangedSettingBool? = null,
+    @field:Json(name = "showArchived") val showArchived: NamedChangedSettingBool? = null,
+    @field:Json(name = "intelligentResumption") val intelligentResumption: NamedChangedSettingBool? = null,
+    @field:Json(name = "autoPlayEnabled") val autoPlayEnabled: NamedChangedSettingBool? = null,
+    @field:Json(name = "hideNotificationOnPause") val hideNotificationOnPause: NamedChangedSettingBool? = null,
+    @field:Json(name = "playUpNextOnTap") val playUpNextOnTap: NamedChangedSettingBool? = null,
+    @field:Json(name = "playOverNotifications") val playOverNotifications: NamedChangedSettingInt? = null,
+    @field:Json(name = "autoUpNextLimit") val autoUpNextLimit: NamedChangedSettingInt? = null,
+    @field:Json(name = "autoUpNextLimitReached") val autoUpNextLimitReached: NamedChangedSettingInt? = null,
+    @field:Json(name = "warnDataUsage") val warnDataUsage: NamedChangedSettingBool? = null,
+    @field:Json(name = "notifications") val showPodcastNotifications: NamedChangedSettingBool? = null,
+    @field:Json(name = "privacyAnalytics") val collectAnalytics: NamedChangedSettingBool? = null,
+    @field:Json(name = "privacyCrashReports") val sendCrashReports: NamedChangedSettingBool? = null,
+    @field:Json(name = "privacyLinkAccount") val linkCrashReportsToUser: NamedChangedSettingBool? = null,
 )
 
 @JsonClass(generateAdapter = true)
@@ -36,6 +60,18 @@ data class NamedChangedSettingInt(
 @JsonClass(generateAdapter = true)
 data class NamedChangedSettingBool(
     @field:Json(name = "value") val value: Boolean,
+    @field:Json(name = "modified_at") val modifiedAt: String,
+)
+
+@JsonClass(generateAdapter = true)
+data class NamedChangedSettingDouble(
+    @field:Json(name = "value") val value: Double,
+    @field:Json(name = "modified_at") val modifiedAt: String,
+)
+
+@JsonClass(generateAdapter = true)
+data class NamedChangedSettingString(
+    @field:Json(name = "value") val value: String,
     @field:Json(name = "modified_at") val modifiedAt: String,
 )